# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freehold API",
    "version": "0.1.0",
    "description": "Governed memory backend for AI agents, built on the Allod format."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Local Freehold"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token"
      }
    },
    "schemas": {
...
      "UpdateEntityBody": {
        "type": "object",
        "properties": {
          "agent": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {}
          },
          "prior": {
            "type": "string",
            "description": "Hash of the changeset being superseded"
          }
        },
        "required": ["agent", "type", "attributes"]
      },
...
      "AdmissionResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["saved", "pending"]
          },
          "hash": {
            "type": "string"
          },
          "proposal": {},
          "rule": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["status", "hash"]
      },
...
      "RecallResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "content": {},
          "author": {
            "type": "string"
          },
          "method": {
            "type": ["string", "null"],
            "description": "Provenance method (model-assisted, manual, etc.) or null for unrecorded"
          },
          "approval": {
            "type": "string",
            "description": "approval field; maps from the `status` query parameter"
          },
          "changeset": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        },
        "required": ["id", "type", "author", "method", "approval", "changeset", "score"]
      },
...
      "MemoryIndexEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "description": "Display title derived from attributes (title, name, statement, first line)"
          },
          "approval": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Taxonomy terms from fold state"
          }
        },
        "required": ["id", "type", "title", "approval", "author", "updatedAt", "terms"]
      },
      "GraphNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "approval": {
            "type": "string"
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "type", "title", "approval", "terms"]
      },
      "GraphEdge": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "required": ["id", "type", "from", "to"]
      },
      "MemoryGraphView": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "True when the node cap cut the listing short"
          }
        },
        "required": ["nodes", "edges", "truncated"]
      }
    },
...
  },
  "paths": {
...
    "/api/v1/entities/{id}": {
      "patch": {
        "summary": "Update an entity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admission result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Entity not found"
          }
        }
      },
      "get": {
        "summary": "Get an entity by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Entity view"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Entity not found"
          }
        }
      }
    },
...
    "/api/v1/memories": {
      "get": {
        "summary": "List recent memories",
        "description": "Most recently indexed memories, newest first. Same result shape as recall (score is 0). The `status` query param maps to the `approval` field. With `scope=all`, returns the full workspace index instead — every non-meta node including pending proposals, as MemoryIndexEntry — and the other query params are ignored.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Set to `all` for the full workspace index (MemoryIndexEntry shape)"
            },
            "required": false,
            "name": "scope",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "author",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Max results (default 50, cap 1000)"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent memories with provenance, or the full index with scope=all",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "anyOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RecallResult"
                          }
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MemoryIndexEntry"
                          }
                        }
                      ]
                    }
                  },
                  "required": ["results"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph": {
      "get": {
        "summary": "Memory graph export",
        "description": "All non-meta nodes and the typed edges between saved nodes, for the graph canvas.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Nodes and edges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryGraphView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/recall": {
      "get": {
        "summary": "Hybrid semantic recall",
        "description": "Semantic + full-text search fused via RRF. The `status` query param maps to the `approval` field in RecallResult.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Search query (required)"
            },
            "required": true,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "author",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by approval state — maps to the `approval` field on RecallResult"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Recall results with provenance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecallResult"
                      }
                    }
                  },
                  "required": ["results"]
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter q"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/entities/{id}/traverse": {
      "get": {
        "summary": "Traverse from an entity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "edgeTypes",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["out", "in", "both"]
            },
            "required": false,
            "name": "direction",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "depth",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Traversal results"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}